+++ next-docs/components/breadcrumbs/Breadcrumbs.tsx
@@ -10,8 +10,8 @@ const Breadcrumbs = () => {
       <a>Home</a>
     </Link>,
     pages.length > 0 &&
-      pages.map((page) => (
-        <Link href={page}>
+      pages.map((page, index) => (
+        <Link href={page} key={index}>
           <a>{page && page[0].toUpperCase() + page.slice(1)}</a>
         </Link>
       )),
@@ -21,7 +21,7 @@ const Breadcrumbs = () => {
   }
   return (
     <div className="relative z-10 hidden lg:block pb-12">
-      <Breadcrumb breadcrumbs={breadcrumbsPath} />
+      <Breadcrumb breadcrumbs={breadcrumbsPath as any} />
     </div>
   );
 };