+++ next-docs/pages/core/insetInput.tsx
@@ -1,0 +1,115 @@
+import React from 'react';
+import Preview from '../../components/codePreview/Preview';
+import ComponentAnatomy from '../../components/ComponentAnatomy';
+import ComponentPageDescription from '../../components/ComponentPageDescription';
+import PropsTable from '../../components/PropsTable';
+import Bg from '../../public/examples/insetInput/Bg';
+import Default from '../../public/examples/insetInput/Default';
+import States from '../../public/examples/insetInput/States';
+import Types from '../../public/examples/insetInput/Types';
+import useExamples from '../../utils/useExamples';
+
+const Example = () => {
+  const examples = useExamples('insetInput');
+  return (
+    <>
+      <ComponentPageDescription title="InsetInput" isInProgress>
+        <p>
+          Text input fields allow users to enter text and can be used to collect
+          user feedback or enter information in data entry forms.
+        </p>
+        <p>
+          These types of input fields are used on their own, or in combination
+          with other inputs such as number entry, date picker, etc.
+        </p>
+      </ComponentPageDescription>
+      <ComponentAnatomy>{`
+    <InsetInput>
+      <InsetInput.Label>...</InsetInput.Label>
+    </InsetInput>
+      `}</ComponentAnatomy>
+      <Preview
+        title="Default"
+        preview={<Default />}
+        code={examples ? examples.Default : 'Loading'}
+      />
+      <Preview
+        title="Different states"
+        preview={<States />}
+        code={examples ? examples.States : 'Loading'}
+      />
+      <Preview
+        title="Custom styles"
+        preview={<Bg />}
+        code={examples ? examples.Bg : 'Loading'}
+      />
+      <Preview
+        title="TextInput types"
+        preview={<Types />}
+        code={examples ? examples.Types : 'Loading'}
+      />
+      <PropsTable
+        title="InsetInput"
+        data={[
+          {
+            name: 'type',
+            type: 'date | datetime-local | email | number | password | search | tel | text | time | url | string',
+            required: false,
+            default: 'text',
+            description: 'Input type',
+          },
+          {
+            name: 'placeholder',
+            type: 'string',
+            required: false,
+            default: '-',
+            description: 'Placeholder for input',
+          },
+          {
+            name: 'error',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: 'Sets error state for input',
+          },
+          {
+            name: 'disabled',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: 'Set disabled/non-disabled',
+          },
+          {
+            name: 'dir',
+            type: 'ltr | rtl | auto',
+            required: false,
+            default: '-',
+            description: 'RTL/LTR direction of label',
+          },
+          {
+            name: 'className',
+            type: 'string',
+            required: false,
+            default: '-',
+            description: 'Tailwind classes for customization',
+          },
+        ]}
+      />
+
+      <PropsTable
+        title="InsetInput.Label"
+        data={[
+          {
+            name: 'className',
+            type: 'string',
+            required: false,
+            default: '-',
+            description: 'Tailwind classes for customization',
+          },
+        ]}
+      />
+    </>
+  );
+};
+
+export default Example;