+++ next-docs/public/examples/input/Default.tsx
@@ -1,0 +1,5 @@
+import { Input } from '@heathmont/moon-core-tw';
+
+const Example = () => <Input />;
+
+export default Example;