+++ next-docs/public/examples/insetInput/Default.tsx
@@ -1,0 +1,9 @@
+import { InsetInput } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <InsetInput placeholder="Placeholder">
+    <InsetInput.Label>Label</InsetInput.Label>
+  </InsetInput>
+);
+
+export default Example;