+++ packages/datepicker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @heathmont/moon-datepicker
 
+## 9.18.1
+
+### Patch Changes
+
+- Bump release
+- Updated dependencies [undefined]
+  - @heathmont/moon-components@9.18.1
+  - @heathmont/moon-core@9.18.1
+  - @heathmont/moon-icons@9.18.1
+  - @heathmont/moon-themes@9.18.1
+  - @heathmont/moon-utils@9.18.1
+
 ## 9.18.0
 
 ### Minor Changes